// src/app/Models/ExceptionElement.model.ts
export class ExceptionElement {
    NonDonors: number;
    CardAmount: number;
    CardDonors: number;
    CashAmount: number;
    CashDonors: number;
    CheckAmount: number;
    CheckDonors: number;
    MailCode: string;
    ID: number;
    DateCaged: Date;
    Source: string;

    showControl: boolean;
    editing: boolean;
    selected: boolean;
}